Backend/requirements.txt
flask
flask-sqlalchemy
pydantic
python-dotenv
google-generativeai
flask-cors
pytest
requests
